+++ apps/commerce-api/src/main/java/com/loopers/application/order/OrderDetailInfo.java
@@ -16,10 +16,10 @@ public record OrderDetailInfo(
     String address,
     String addressDetail,
     String shippingMemo,
-    int totalAmount,
-    int shippingFee,
-    int discountAmount,
-    int paymentAmount,
+    Long totalAmount,
+    Long shippingFee,
+    Long discountAmount,
+    Long paymentAmount,
     List<OrderProductInfo> orderProducts
 ) {
