+++ apps/commerce-api/src/main/java/com/loopers/application/order/OrderInfo.java
@@ -8,8 +8,8 @@ public record OrderInfo(
     String orderNumber,
     String orderName,
     OrderStatus status,
-    int totalAmount,
-    int paymentAmount,
+    Long totalAmount,
+    Long paymentAmount,
     String thumbnailUrl
 ) {
